Adding confirmations

+++ app/components/notification-channel-list.js
@@ -18,6 +18,11 @@ export default class NotificationChannelListComponent extends Component {
 
   @action
   async deleteChannel(channel) {
+    if (
+      !confirm('Are you sure you want to delete this notification channel?')
+    ) {
+      return;
+    }
     await channel.destroyRecord();
     set(this, 'channels', await this.store.findAll('notification-channel'));
   }
